+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Update `NumberFormat.TOKEN_AMOUNT_SHORT` to round small numbers (between -0.01 and 0.01) to the nearest hundredth
+    - Add configurable options (`smallValueThreshold` and `smallValueFractionDigits`) to INumberFormat.
+    - Update `TOKEN_AMOUNT_SHORT` defaults in numberFormats.
 
 ## [1.0.68] - 2025-02-21